Handle older output format
This commit updates the restore format to handle an older output format
where each folder was associated with an array of strings that were the
UUIDs of the VMs in the folder.

The `dump-vcenter-folders.rb` script produces an array of hashes that
includes more info, such as the VM name.