Add ability to exclude paths from dump or restore
This commit adds the ability to exclude VMs and folders by path during both
dump and restore. The feature is implemented in terms of the `-W` and
`--exclude` flags (borrowed from TAR) where each use of the flag is followed by
a glob string. The glob syntax used is that understood by Ruby's File.fnmatch
function with the FNM_PATHNAME and FNM_EXTGLOB flags set. This is roughly
equivalent to Zsh or Bash v4 globs.

A VM or folder will be excluded from a dump or restore operation
if its _full path_ matches any of the specified globs.

For example, to exclude a directory:

    ./dump-vcenter-folders.rb ... --exclude 'root-dir/dir-to-exclude'

Or any entry with a filename that starts with "foovm-":

    ./restore-vcenter-folders.rb ... --exclude '**/foovm-*'

Closes Sharpie/vsphere-folder-copy#4.